pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "auto-research"
version = "0.1.0"
description = "Auto research pipeline: experiments, feedback, CI/CD hooks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4.0"]
anthropic = ["anthropic>=0.40.0"]

[project.scripts]
auto-research = "auto_research.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/auto_research"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
